# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Lightning Talk Event Management API
  description: |
    ライトニングトークイベント管理システムのREST API
    
    このAPIは以下の機能を提供します：
    - イベント管理 (作成、編集、削除)
    - 参加者管理 (登録、編集、リスト表示)
    - 発表管理 (提出、承認、編集)
    - リアルタイム通知 (WebSocket, Server-Sent Events)
    - 複数イベント管理 (並行実行、競合解決)
    - 分析レポート (統計、トレンド分析)
    - 管理者機能 (ダッシュボード、設定)
  version: 1.0.0
  contact:
    name: Lightning Talk Support
    email: support@lightningtalk.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: 開発環境
  - url: https://api.lightningtalk.example.com/api
    description: 本番環境
paths:
...
  /multi-events/create-batch:
    post:
      tags:
        - Multi-Events
      summary: 複数イベント一括作成
      description: 複数のイベントを一括で作成します
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - events
              properties:
                events:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/EventCreateRequest'
                options:
                  type: object
                  properties:
                    checkConflicts:
                      type: boolean
                      default: true
                    autoResolve:
                      type: boolean
                      default: false
                    notifyParticipants:
                      type: boolean
                      default: true
                    batchSize:
                      type: integer
                      minimum: 1
                      maximum: 50
                      default: 10
      responses:
        '201':
          description: 一括作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "5件のイベントが正常に作成されました"
                  data:
                    type: object
                    properties:
                      created:
                        type: array
                        items:
                          $ref: '#/components/schemas/Event'
                      conflicts:
                        type: array
                        items:
                          type: object
                      errors:
                        type: array
                        items:
                          type: object
                      summary:
                        type: object
                        properties:
                          totalAttempted:
                            type: integer
                          successfullyCreated:
                            type: integer
                          conflictsFound:
                            type: integer
                          errorsOccurred:
                            type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /multi-events/concurrent:
    get:
      tags:
        - Multi-Events
      summary: 並行イベント管理
      description: 並行実行中のイベントの管理情報を取得します
      security:
        - AdminAuth: []
      parameters:
        - name: dateRange.start
          in: query
          schema:
            type: string
            format: date-time
        - name: dateRange.end
          in: query
          schema:
            type: string
            format: date-time
        - name: includeConflicts
          in: query
          schema:
            type: boolean
            default: true
        - name: resolveConflicts
          in: query
          schema:
            type: boolean
            default: false
        - name: priorityOrder
          in: query
          schema:
            type: string
            enum: [date, priority, participants]
            default: date
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "並行イベント管理情報を取得しました"
                  data:
                    type: object
                    properties:
                      activeEvents:
                        type: integer
                      concurrentGroups:
                        type: array
                      conflicts:
                        type: object
                      managementStatus:
                        type: object
                        properties:
                          healthy:
                            type: boolean
                          warnings:
                            type: array
                            items:
                              type: string
                          recommendations:
                            type: array
                            items:
                              type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /multi-events/dashboard:
    get:
      tags:
        - Multi-Events
      summary: リアルタイムダッシュボード
      description: マルチイベント管理のリアルタイムダッシュボードデータを取得します
      security:
        - AdminAuth: []
      parameters:
        - name: refreshInterval
          in: query
          schema:
            type: integer
            minimum: 5
            maximum: 300
            default: 30
        - name: includeMetrics
          in: query
          schema:
            type: boolean
            default: true
        - name: includeAlerts
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # 分析レポート
...
components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: string
          example: "event-001"
        title:
          type: string
          example: "第1回 なんでもライトニングトーク"
        description:
          type: string
          example: "5分間で世界を変える！あなたの「なんでも」を聞かせて！"
        date:
          type: string
          format: date-time
          example: "2025-06-25T19:00:00+09:00"
        endDate:
          type: string
          format: date-time
          example: "2025-06-25T22:00:00+09:00"
        venue:
          type: object
          properties:
            name:
              type: string
              example: "新宿某所"
            address:
              type: string
              example: "詳細は6月20日に確定予定"
            capacity:
              type: integer
              example: 50
            online:
              type: boolean
              example: true
            onlineUrl:
              type: string
              example: "https://meet.google.com/ycp-sdec-xsr"
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
          example: "upcoming"
        registrationOpen:
          type: boolean
          example: true
        talkSubmissionOpen:
          type: boolean
          example: true
        maxTalks:
          type: integer
          example: 20
        talkDuration:
          type: integer
          example: 5
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventCreateRequest:
      type: object
      required:
        - title
        - description
        - date
        - venue
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "テストイベント"
        description:
          type: string
          minLength: 1
          maxLength: 1000
          example: "テスト用のイベントです"
        date:
          type: string
          format: date-time
          example: "2025-12-01T19:00:00Z"
        endDate:
          type: string
          format: date-time
          example: "2025-12-01T22:00:00Z"
        venue:
          type: object
          required:
            - name
          properties:
            name:
              type: string
              example: "テスト会場"
            address:
              type: string
              example: "東京都新宿区"
            capacity:
              type: integer
              minimum: 1
              example: 30
            online:
              type: boolean
              default: false
            onlineUrl:
              type: string
              example: "https://meet.google.com/test"
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
          default: upcoming
        registrationOpen:
          type: boolean
          default: true
        talkSubmissionOpen:
          type: boolean
          default: true
        maxTalks:
          type: integer
          minimum: 1
          default: 20
        talkDuration:
          type: integer
          minimum: 1
          default: 5
...
    DashboardData:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        refreshInterval:
          type: integer
        overview:
          type: object
          properties:
            activeEvents:
              type: integer
            concurrentGroups:
              type: integer
            totalConflicts:
              type: integer
            crossEventParticipants:
              type: integer
        metrics:
          type: object
          properties:
            eventUtilization:
              type: number
              minimum: 0
              maximum: 1
            resourceEfficiency:
              type: number
              minimum: 0
              maximum: 1
            participantSatisfaction:
              type: number
              minimum: 0
              maximum: 1
        alerts:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [warning, error, info]
              message:
                type: string
              action:
                type: string
        recommendations:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [optimization, insight, action]
              message:
                type: string
              priority:
                type: string
                enum: [high, medium, low]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "エラーが発生しました"
        error:
          type: string
          example: "詳細エラー情報"
        timestamp:
          type: string
          format: date-time
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "バリデーションエラー"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "有効なメールアドレスを入力してください"
              value:
                type: string
                example: "invalid-email"
  responses:
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "認証が必要です"
...
    InternalServerError:
      description: サーバー内部エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 管理者用JWTトークン

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: APIキー認証
...